requirements.txt
fastapi
uvicorn
pydantic
transformers
diffusers
python-jose[cryptography]
passlib
python-multipart
bcrypt==4.0.1
pytest
requests
pillow
pytest-asyncio
httpx
huggingface_hub[cli]
accelerate
peft
protobuf
sentencepiece
bitsandbytes
